StudentApp/.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
google-cloud-aiplatform==1.38.0
pandas==2.0.3
fsspec==2023.6.0
gcsfs==2023.6.0
db-dtypes==1.1.1